backend/requirements.txt
fastapi
pydantic
motor
passlib
python-jose
python-dotenv
